+++ Hardware/PWM.h
@@ -2,7 +2,7 @@
 #define __PWM_H
 
 void PWM_Init(void);
-void PWM_SetCompare1(uint16_t Compare1);
-void PWM_SetCompare2(uint16_t Compare2);
+void PWM_SetCompare1(uint8_t Compare1);
+void PWM_SetCompare2(uint8_t Compare2);
 
 #endif
